Add Cosmos examples -Add cosmos decorators -Fix input params -Add Import React to works with cosmos -Fix imports -Rename TextInput
