+++ src/Components/Form/InputText/index.tsx
@@ -1,1 +1,1 @@
-export { default } from "./text-input";
+export { default } from "./textInput";
