+++ src/__fixtures__/label.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+import Label from "../Components/Form/Label";
+
+export default <Label value="Label" />;
